Longer commit

+++ src/Main.java
@@ -1,7 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {2, 3, 4, 5};
+        long[] sales = {2, 3, 4, 5};
         SalesManager salesManager = new SalesManager(sales);
         System.out.println(salesManager.max());
     }
 }
+
